move to examples; add argparse
